requirements.txt
# Web Framework
nicegui==1.4.21
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# UUID
uuid==1.30

# HTTP Client (for SIEM webhooks)
httpx==0.26.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3